Replace to unions, replace to uint32_t, add LUTs

+++ firmware/cartridges/irq.h
@@ -6,7 +6,8 @@
 void irq_handler(void);
 void initPorts(void);
 void reset_SID(void);
-void setreg(uint8_t addr,uint8_t data);
+//uint32_t instead of uint8_t allows to compile without UXTB instruction, its not needed
+void setreg(uint32_t addr,uint32_t data);
 void InitHardware(void);
